udatapath: use C++ linker
The nbee library is written in C++ so we have to use the C++ linker. Normally
automake detects which linker to use from the source file extensions, but
all of our sources are C. The trick suggested in the automake documentation [1]
is to add a fake C++ source file which doesn't actually need to exist.

Fixes #24.

[1]: http://www.gnu.org/software/automake/manual/html_node/Libtool-Convenience-Libraries.html