Re-written match_std.c module:
Fixed several faults in matching of packets against flow entries and
matching two flow entries both strictly and non-strictly.

Removed duplicated functions and aligned the different match functions
as much as possible to simplify the code and highlight the differences.

New implementation of match_std_overlap(), which was broken already in
of11softswitch