Create dpctl_parser.py
A simple python utility which helps monitoring and debugging 
the flows set at the switch (using OF1.3). It displays the flows per 
table in a more convinient way. 
It uses the dpctl tool to send the flow stats req.
 
To run the application simply supply the datapath used, for example:

 pyhon2.7 dpctl_parser.py unix:/var/run/dp0