Fix strict masked ip matching.
Quick and dirty hack to stop different IP addresses, with equal masks,
to match.
The problem is caused because we work with IP addresses in the network
byte order format and the function strict_match32 performs matching
on values in the host byte order. Thus, in little endian machines,
the function will not work.
Problems like this show  that the match needs a
huge refactoring.