Minor refactor to meter token bucket.
This commit modified the token bucket capacity unity.
Previously the code considered tokens in Kbps. For this
reason the packet size, which is presented in bytes,
value was also converted. This packet size convertion was wrong,
because small packets sizes were truncated to 0 in Kbps preventing
token remotion.

The code now changes the bucket capacity and packet conversion to bits.
It avoids divisions and eliminates errors of truncation.

Conflicts:
	udatapath/meter_entry.c