Copy packet to packet_in buffer
In the BEBA code the packet is allocated in the stack.
When the packet has to be sent to the control plane and
buffering is in place, the packet must be copied to the
buffer rather than only a reference. Otherwise, the packet
is released after the pipeline is complete, which causes
a segmentation fault when trying to retrieve and use
the packet from the buffer.